+++ p32_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 )
 
-func p32NumDigits(a int) int {
+func p32NumDigits(a int64) int {
 	switch {
 	case a < 10:
 		return 1
@@ -35,10 +35,10 @@ func p32NumDigits(a int) int {
 }
 
 // Return c=a*b if "a*b=c" is pandigital; return 0 otherwise
-func p32PanDigitalProduct(a, b int) int {
+func p32PanDigitalProduct(a, b int64) int64 {
 	c := a * b
 
-	if natural.PandigitalMulti([]int{a, b, c}, 9, false) {
+	if natural.PandigitalMulti([]int64{a, b, c}, 9, false) {
 		return c
 	}
 	return 0
@@ -50,15 +50,15 @@ func TestP32(t *testing.T) {
 		t.Fail()
 	}
 
-	pandig := make(map[int]interface{})
+	pandig := make(map[int64]interface{})
 
-	sum := 0
-	for a := 2; a < 999; a++ {
+	sum := int64(0)
+	for a := int64(2); a < 999; a++ {
 		ad := p32NumDigits(a)
 		bdmin := (9 - ad) / 2
 		bmin := intgr.Pow(10, bdmin-1)
 		for b := bmin; b < bmin*10; b++ {
-			c := p32PanDigitalProduct(a, b)
+			c := p32PanDigitalProduct(a, int64(b))
 			if c != 0 {
 				if _, found := pandig[c]; !found {
 					sum += c
