+++ p38_test.go
@@ -6,13 +6,13 @@ import (
 )
 
 func TestP38(t *testing.T) {
-	max := 0
-	for i := 1; i < 99999; i++ {
-		j := 1
-		prod := 0
+	max := int64(0)
+	for i := int64(1); i < 99999; i++ {
+		j := int64(1)
+		prod := int64(0)
 		proddig := make([]byte, 0, 9)
 		for len(proddig) < 9 {
-			prod = natural.Concat([]int{prod, i * j})
+			prod = natural.Concat([]int64{prod, i * j})
 			proddig = append(proddig, natural.Digits(i*j)...)
 			j += 1
 		}
